test(tavily): split free account boundary coverage
